remove duplicate implementation of contentservice
